HT-3003: missing largeclusters file bugfix  * WARN not ERROR  * Empty set of OCNS

+++ lib/data_sources/large_clusters.rb
@@ -16,6 +16,10 @@ module DataSources
     def load_large_clusters
       ocns = Set.new
       Services.logger.info("Loading large clusters file #{@filename}")
+      unless File.exist? Settings.large_cluster_ocns
+        Services.logger.warn("No large clusters file found.")
+        return ocns
+      end
       File.open(Settings.large_cluster_ocns).each_line do |line|
         ocns.add(line.to_i)
       end
